+++ packages/prism_theme_flutter/lib/prism_theme_flutter.dart
@@ -6,5 +6,7 @@ export 'package:prism_flutter/prism_flutter.dart';
 export 'package:prism_theme/prism_theme.dart';
 
 export 'src/converters.dart';
+export 'src/focus_ring.dart';
+export 'src/text.dart';
 export 'src/theme_data.dart';
 export 'src/theme_extension.dart';